Add package pieces and move classes to it
