+++ src/main/java/net/honux/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package net.honux.chess;
+package net.honux.pieces;
 
 public class Pawn {
